chore: refactor to support multiple cases
