+++ frontend/src/app/components/containers/cases/case-activities/caseActivityCard.tsx
@@ -11,13 +11,13 @@ import { CODE_TABLE_TYPES } from "@/app/constants/code-table-types";
 import { complaintTypeToName } from "@/app/types/app/complaint-types";
 import Option from "@apptypes/app/option";
 
-interface CaseActivitiesProps {
-  caseGuid: string;
+interface CaseActivityCardProps {
+  caseIdentifier: string;
   caseName?: string;
 }
 
-export const CaseActivities: FC<CaseActivitiesProps> = ({ caseGuid, caseName }) => {
-  const { linkedComplaints, investigations, inspections, isLoading } = useCaseActivities(caseGuid);
+export const CaseActivityCard: FC<CaseActivityCardProps> = ({ caseIdentifier, caseName }) => {
+  const { linkedComplaints, investigations, inspections, isLoading } = useCaseActivities(caseIdentifier);
 
   const agencies = useAppSelector(selectCodeTable(CODE_TABLE_TYPES.AGENCY));
   const complaintStatusCodes = useAppSelector(selectCodeTable(CODE_TABLE_TYPES.COMPLAINT_STATUS));
@@ -162,67 +162,63 @@ export const CaseActivities: FC<CaseActivitiesProps> = ({ caseGuid, caseName }) 
   };
 
   return (
-    <>
-      <div className="border rounded p-4 mb-4 bg-white">
-        <div className="d-flex flex-column mb-3">
-          <span className="mb-2 mb-sm-0 larger-font">
-            <b>Case</b>&nbsp;
-            <Link
-              to={`/case/${caseGuid}`}
-              className="comp-cell-link fw-semibold"
-            >
-              #{caseName ?? caseGuid}
-            </Link>
-          </span>
-        </div>
-
-        {isLoading ? (
-          <div className="d-flex align-items-center gap-2 text-muted">
-            <Spinner
-              animation="border"
-              size="sm"
-            />
-            <span>Loading associated data...</span>
-          </div>
-        ) : (
-          <>
-            {linkedComplaints && linkedComplaints.length > 0 && (
-              <div className="mb-3 ms-3">
-                <div className="mb-2 text-black-50">
-                  <i className="bi bi-arrow-return-right text-black"></i>&nbsp;Complaint(s):
-                </div>
-                <ul className="mb-0 ms-4 list-unstyled">
-                  {linkedComplaints.map((complaint) => renderComplaintItem(complaint))}
-                </ul>
-              </div>
-            )}
-
-            {investigations && investigations.length > 0 && (
-              <div className="mb-3 ms-3">
-                <div className="mb-2 text-black-50">
-                  <i className="bi bi-arrow-return-right text-black"></i>&nbsp;Investigation(s):
-                </div>
-                <ul className="mb-0 ms-4 list-unstyled">
-                  {investigations.map((investigation) => renderInvestigationItem(investigation))}
-                </ul>
-              </div>
-            )}
-
-            {inspections && inspections.length > 0 && (
-              <div className="mb-3 ms-3">
-                <div className="mb-2 text-black-50">
-                  <i className="bi bi-arrow-return-right text-black"></i>&nbsp;Inspection(s):
-                </div>
-                <ul className="mb-0 ms-4 list-unstyled">
-                  {inspections.map((inspection) => renderInspectionItem(inspection))}
-                </ul>
-              </div>
-            )}
-          </>
-        )}
+    <div className="border rounded p-4 mb-4 bg-white">
+      <div className="d-flex flex-column mb-3">
+        <span className="mb-2 mb-sm-0 larger-font">
+          <b>Case</b>&nbsp;
+          <Link
+            to={`/case/${caseIdentifier}`}
+            className="comp-cell-link fw-semibold"
+          >
+            #{caseName ?? caseIdentifier}
+          </Link>
+        </span>
       </div>
-    </>
+
+      {isLoading ? (
+        <div className="d-flex align-items-center gap-2 text-muted">
+          <Spinner
+            animation="border"
+            size="sm"
+          />
+          <span>Loading associated data...</span>
+        </div>
+      ) : (
+        <>
+          {linkedComplaints && linkedComplaints.length > 0 && (
+            <div className="mb-3 ms-3">
+              <div className="mb-2 text-black-50">
+                <i className="bi bi-arrow-return-right text-black"></i>&nbsp;Complaint(s):
+              </div>
+              <ul className="mb-0 ms-4 list-unstyled">
+                {linkedComplaints.map((complaint) => renderComplaintItem(complaint))}
+              </ul>
+            </div>
+          )}
+
+          {investigations && investigations.length > 0 && (
+            <div className="mb-3 ms-3">
+              <div className="mb-2 text-black-50">
+                <i className="bi bi-arrow-return-right text-black"></i>&nbsp;Investigation(s):
+              </div>
+              <ul className="mb-0 ms-4 list-unstyled">
+                {investigations.map((investigation) => renderInvestigationItem(investigation))}
+              </ul>
+            </div>
+          )}
+
+          {inspections && inspections.length > 0 && (
+            <div className="mb-3 ms-3">
+              <div className="mb-2 text-black-50">
+                <i className="bi bi-arrow-return-right text-black"></i>&nbsp;Inspection(s):
+              </div>
+              <ul className="mb-0 ms-4 list-unstyled">
+                {inspections.map((inspection) => renderInspectionItem(inspection))}
+              </ul>
+            </div>
+          )}
+        </>
+      )}
+    </div>
   );
 };
-
-export default CaseActivities;
